refactor: move ConfirmDeleteTask to presentation.deletetask, use CancelableActionLayout shared component and change CustomDragHandle color
