+++ app/src/main/java/com/washingtondcsquad/tudee/presentation/deletetask/ConfirmDeleteTask.kt
@@ -1,4 +1,4 @@
-package com.washingtondcsquad.tudee.presentation.components.deletetask
+package com.washingtondcsquad.tudee.presentation.deletetask
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
@@ -14,8 +14,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
@@ -25,15 +23,18 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.washingtondcsquad.tudee.R
-import com.washingtondcsquad.tudee.presentation.design.theme.AppTheme
+import com.washingtondcsquad.tudee.presentation.components.CancelableActionLayout
+import com.washingtondcsquad.tudee.presentation.design.AppTheme
 
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -47,20 +48,20 @@ fun ConfirmDeleteTask(
 
 ) {
 
-    val sheetState=rememberModalBottomSheetState()
+    val sheetState = rememberModalBottomSheetState()
     LaunchedEffect(Unit) {
         sheetState.show()
     }
 
     ModalBottomSheet(
-        onDismissRequest = {onDismiss()},
+        onDismissRequest = { onDismiss() },
         modifier = Modifier
             .fillMaxWidth(),
         shape = RoundedCornerShape(topEnd = 24.dp, topStart = 24.dp),
         containerColor = AppTheme.colors.surface,
         sheetState = sheetState,
-        dragHandle =null
-        ) {
+        dragHandle = null
+    ) {
 
         Column(
             modifier = Modifier.fillMaxWidth(),
@@ -85,30 +86,18 @@ fun ConfirmDeleteTask(
             Spacer(modifier = Modifier.padding(bottom = 24.dp))
 
             // blur
-            Column(
+            CancelableActionLayout(
+                actionText = stringResource(R.string.delete),
+                actionTextColor = AppTheme.colors.error,
+                onAction = {deleteOnClick()},
+                onCancel = {cancelOnClick()},
                 modifier = Modifier
                     .fillMaxWidth()
                     .height(148.dp)
-                    .shadow(elevation = 20.dp, spotColor = Color.Black.copy(alpha = 0.08f))
-                    .border(BorderStroke(width = 1.dp, AppTheme.colors.surfaceHigh))
-                    .padding(vertical = 12.dp, horizontal = 16.dp),
-                verticalArrangement = Arrangement.spacedBy(10.dp)
-
-            ) {
-                DeleteButton(
-                    buttonColor = AppTheme.colors.errorVariant,
-                    buttonTextId = R.string.delete,
-                    buttonTextColor = AppTheme.colors.error,
-                    buttonClick = deleteOnClick
-                )
-                DeleteButton(
-                    borderStroke = 1,
-                    borderColor = AppTheme.colors.stroke,
-                    buttonTextId = R.string.cancel,
-                    buttonTextColor = AppTheme.colors.primary,
-                    buttonClick = cancelOnClick
-                )
-            }
+                    .shadow(elevation = 20.dp, spotColor = Color.Black.copy(alpha = 0.08f)),
+                actionBackgroundColor = List(2) { AppTheme.colors.errorVariant },
+                isEnabled = true
+            )
         }
     }
 
@@ -156,47 +145,7 @@ fun DeleteTaskIllustration() {
 }
 
 @Composable
-fun DeleteButton(
-    borderStroke: Int = 0,
-    borderColor: Color = Color.Transparent,
-    buttonColor: Color = AppTheme.colors.surfaceHigh,
-    buttonTextId: Int,
-    buttonTextColor: Color,
-    buttonClick: () -> Unit
-) {
-
-    Button(
-        onClick = { /*buttonClick()*/ },
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(58.dp),
-        shape = RoundedCornerShape(100.dp),
-        colors = ButtonDefaults.buttonColors(buttonColor),
-        elevation = null,
-        border = BorderStroke(width = borderStroke.dp, borderColor)
-
-    ) {
-
-        Row(
-            modifier = Modifier.padding(vertical = 8.dp, horizontal = 24.dp),
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-
-                text = stringResource(buttonTextId),
-                style = AppTheme.textStyle.label.large,
-                color = buttonTextColor,
-
-                )
-        }
-
-
-    }
-}
-
-@Composable
-fun CustomDragHandle(modifier: Modifier=Modifier){
+fun CustomDragHandle(modifier: Modifier = Modifier) {
     Row(
         modifier = modifier.fillMaxWidth(),
         horizontalArrangement = Arrangement.Center
@@ -205,7 +154,7 @@ fun CustomDragHandle(modifier: Modifier=Modifier){
             modifier = Modifier
                 .size(width = 32.dp, height = 4.dp)
                 .clip(RoundedCornerShape(100))
-                .background(AppTheme.colors.body),
+                .background(AppTheme.colors.body.copy(alpha = AppTheme.colors.body.alpha * 0.4f))
         )
     }
 }
